+++ src/Sulu/Component/Content/Tests/Unit/SmartContent/ContentDataProviderTest.php
@@ -227,7 +227,6 @@ class ContentDataProviderTest extends \PHPUnit_Framework_TestCase
             ['uuid' => '123-123-789', 'title' => 'My-Page-2', 'path' => '/my-page-2'],
         ];
 
-        $referenceStore = new ReferenceStore();
         $provider = new ContentDataProvider(
             $this->getContentQueryBuilder(
                 [
@@ -241,7 +240,7 @@ class ContentDataProviderTest extends \PHPUnit_Framework_TestCase
             $this->getDocumentManager(['123-123-123' => $data[0], '123-123-456' => $data[1]]),
             $this->getProxyFactory(),
             $this->getSession(),
-            $referenceStore,
+            new ReferenceStore(),
             true
         );
 
@@ -261,7 +260,6 @@ class ContentDataProviderTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals($data[1]['uuid'], $items[1]->getId());
         $this->assertEquals($data[1], $items[1]->getResource()->getWrappedValueHolderValue());
         $this->assertTrue($result->getHasNextPage());
-        $this->assertEquals(['123-123-123', '123-123-456'], $referenceStore->getAll());
     }
 
     public function testResolveDataItemsOnlyPublished()
@@ -315,6 +313,7 @@ class ContentDataProviderTest extends \PHPUnit_Framework_TestCase
             ['uuid' => '123-123-789', 'title' => 'My-Page-2', 'path' => '/my-page-2'],
         ];
 
+        $referenceStore = new ReferenceStore();
         $provider = new ContentDataProvider(
             $this->getContentQueryBuilder(
                 [
@@ -328,7 +327,7 @@ class ContentDataProviderTest extends \PHPUnit_Framework_TestCase
             $this->getDocumentManager(['123-123-123' => $data[0], '123-123-456' => $data[1]]),
             $this->getProxyFactory(),
             $this->getSession(),
-            new ReferenceStore(),
+            $referenceStore,
             true
         );
 
@@ -348,6 +347,7 @@ class ContentDataProviderTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals($data[1]['uuid'], $items[1]->getId());
         $this->assertEquals($data[1], $items[1]->getResource()->getWrappedValueHolderValue());
         $this->assertTrue($result->getHasNextPage());
+        $this->assertEquals(['123-123-123', '123-123-456'], $referenceStore->getAll());
     }
 
     public function testResolveDataItemsNoPagination()